feat: 닉네임 input 컴포넌트 분리 및 마이페이지 적용
